refactor: remove obsolete accordion content block in separate file
