+++ admin/src/common/blocks/AccordionItemBlock.tsx
@@ -1,8 +1,28 @@
 import { SwitchField } from "@comet/admin";
-import { BlockCategory, BlocksFinalForm, createCompositeBlock, createCompositeBlockTextField, createCompositeSetting } from "@comet/blocks-admin";
+import {
+    BlockCategory,
+    BlocksFinalForm,
+    createBlocksBlock,
+    createCompositeBlock,
+    createCompositeBlockTextField,
+    createCompositeSetting,
+} from "@comet/blocks-admin";
 import { AccordionItemBlockData } from "@src/blocks.generated";
-import { AccordionContentBlock } from "@src/common/blocks/AccordionContentBlock";
+import { CallToActionListBlock } from "@src/common/blocks/CallToActionListBlock";
+import { HeadingBlock } from "@src/common/blocks/HeadingBlock";
+import { RichTextBlock } from "@src/common/blocks/RichTextBlock";
+import { SpaceBlock } from "@src/common/blocks/SpaceBlock";
 import { FormattedMessage } from "react-intl";
+
+const AccordionContentBlock = createBlocksBlock({
+    name: "AccordionContent",
+    supportedBlocks: {
+        richtext: RichTextBlock,
+        space: SpaceBlock,
+        heading: HeadingBlock,
+        callToActionList: CallToActionListBlock,
+    },
+});
 
 export const AccordionItemBlock = createCompositeBlock(
     {
