--- site/src/common/blocks/AccordionContentBlock.tsx
@@ -1,20 +1,0 @@
-import { BlocksBlock, PropsWithData, SupportedBlocks, withPreview } from "@comet/cms-site";
-import { AccordionContentBlockData } from "@src/blocks.generated";
-import { CallToActionListBlock } from "@src/common/blocks/CallToActionListBlock";
-import { HeadingBlock } from "@src/common/blocks/HeadingBlock";
-import { RichTextBlock } from "@src/common/blocks/RichTextBlock";
-import { SpaceBlock } from "@src/common/blocks/SpaceBlock";
-
-const supportedBlocks: SupportedBlocks = {
-    richtext: (props) => <RichTextBlock data={props} />,
-    heading: (props) => <HeadingBlock data={props} />,
-    space: (props) => <SpaceBlock data={props} />,
-    callToActionList: (props) => <CallToActionListBlock data={props} />,
-};
-
-export const AccordionContentBlock = withPreview(
-    ({ data: { blocks } }: PropsWithData<AccordionContentBlockData>) => {
-        return <BlocksBlock data={{ blocks }} supportedBlocks={supportedBlocks} />;
-    },
-    { label: "Accordion Content" },
-);
